# perf: index deployments by risk score for main dashboard