### Description

<!--
A detailed explanation of the changes in your PR. Feel free to remove this
section if the title of your PR is sufficiently descriptive. To learn more
about contributing to this project, check "*.md" files under:
    https://github.com/stackrox/stackrox/tree/master/.github
-->

Add index on deployments to improve main dashboard performances when users have restricted access to a large number of namespaces. There was previous work on the performance topic, but this solution, pointed out in #13759 is simpler and about as efficient.

Benchmark results (elapsed in ms of /v1/deployments?pagination.sortOption.field=Deployment%20Risk%20Priority)

| scope size (#ns) | 4.5.5 | riskScore index |
|---|---|---|
| unrestricted | 28 ms | 5 ms |
| 1 | 25 ms | 25 ms |
| 10 | 25 ms | 25 ms |
| 20 | 26 ms | 26 ms |
| 50 | 29 ms | 29 ms |
| 100 | 33 ms | 30 ms |
| 200 | 43 ms | 34 ms |
| 500 | 174 ms | 42 ms |
| 1000 | 498 ms | 65 ms |
| 2000 | 907 ms | 84 ms |

### User-facing documentation

- [x] CHANGELOG ~~is updated **OR**~~ update is not needed
- [x] [documentation PR](https://spaces.redhat.com/display/StackRox/Submitting+a+User+Documentation+Pull+Request) ~~is created and is linked above **OR**~~ is not needed

### Testing and quality

<!--
General Availability requirements: https://github.com/stackrox/stackrox/blob/master/PR_GA.md
Feature Flags usage: https://github.com/stackrox/stackrox/blob/master/pkg/features/README.md
-->

- [x] the change is production ready: the change is GA or otherwise the functionality is gated by a feature flag
- [ ] CI results are inspected

#### Automated testing

<!--
If no tests have been contributed, please explain why unless it's obvious,
e.g., the PR is a one-line comment change.

- [ ] added unit tests
- [ ] added e2e tests
- [ ] added regression tests
- [ ] added compatibility tests
- [ ] modified existing tests
-->

#### How I validated my change

<!--
Use this space to explain **how you validated** that **your change functions
exactly how you expect it**. Feel free to attach JSON snippets, curl commands,
screenshots, etc. Apply a simple benchmark: would the information you provided
convince any reviewer or any external reader that you did enough to validate
your change.

It is acceptable to assume trust and keep this section light, e.g. as a
bullet-point list.

It is acceptable to skip testing in cases when CI is sufficient, or it's a
markdown or code comment change only. It is also acceptable to skip testing for
changes that are too taxing to test before merging. In such case you are
responsible for the change after it gets merged which includes reverting,
fixing, etc. Make sure you validate the change ASAP after it gets merged or
explain in PR when the validation will be performed. Explain here why you
skipped testing in case you did so.

Have you created automated tests for your change? Explain here which validation
activities you did manually and why so.
-->

run altered tests/performance/load